Move external blueprint declaration to the view
Moves external views to a separate app directory and makes the
file a bit more self-contained by including the blueprint
declaration in the view file itself.
